call_notes_app/agentcore_agent/requirements.txt
bedrock-agentcore
strands-agents
mcp
boto3>=1.35.0
uv
